Further split `types/mod.rs`
